requirements.txt
python-telegram-bot==21.0.1
python-dotenv==1.0.1
openai==1.54.3
httpx==0.27.2
aiofiles==24.1.0

# Optional: Redis support for bot communication (production)
# redis==5.0.1